Incorporate Shiftable and pass respective enigma_spec tests
